Require a valid connection when quoting a string
The underlying Trilogy driver will raise an IOError inside #escape if
the socket is closed, so we need to ensure we have a valid connection,
and allow retries if necessary, when quoting a string.

IOError was also not previously on the list of known connectivity-
related errors, so fix that too, by treating it the same as SocketError.

Co-authored-by: John Hawthorn <john@hawthorn.email>