Support Rails 5.2, 6.x, and 7.0
Seeing the performance advantages of the Trilogy driver compared
with mysql2 has motivated me to make it compatible with older Rails
versions. I'm hoping it can become the MySQL driver of choice in
the Rails community. After trying several approaches, have ended up
with the solution you see in this PR.

**ActiveRecord Trilogy Adapter** was originally written with a
forward-thinking mind set -- one exclusive to Rails 7.1 and later.
The older database adapter framework found in prior versions of
Rails differs a fair bit from that found in 7.1, so there are
some core differences to bridge in order to have Trilogy become
compatible with prior Rails versions. This PR adds backwards
compatibility in such a way that it does not interfere with the
lean nature of how AR Trilogy has been architected -- it adds
that compatibility layer via a series of patches all tucked away in
a separate file. Inside that compatibility file,
backwards_ar_compatibility.rb, checks are done to identify which
newer methods are missing. They are selectively added back in, with
the ultimate goal of establishing the broadest compatibility
possible. Only those patches which are necessary for any given
Rails version get applied, making it minimally invasive.

This approach requires no changes to Trilogy itself, so it
functions in exactly the same way as it always has done -- as if
everything is just ActiveRecord 7.1+. As such newer Rails 7.1 apps
are completely unaffected because for them none of that patching
needs be applied.