Remove translation of exception on reconnect
This translation logic breaks Rails parallel tests, because that
explicitly rescues the underlying ActiveRecord::NoDatabaseError. This
means it fails all tests when running with parallel tests.

Based on Rails upstream, this translation is done there either so this
drops it here as well. One test had to be slightly updated for this. But
that test is also deleted already on upstream Rails.